Consolidate XML Schema collision tests and fixtures
