--- tests/data/expected/main/xmlschema/field_name_collisions/default_namespace_attributes.txt
@@ -1,1 +1,0 @@
-XML Schema cannot represent both attribute '{urn:one}code' and attribute '{urn:two}code' as property 'code'.
