--- tests/data/expected/main/xmlschema/field_name_collisions/default_namespace_distinct.txt
@@ -1,1 +1,0 @@
-XML Schema cannot represent both element '{urn:one}code' and element '{urn:two}code' as property 'code'.
