--- tests/data/expected/main/xmlschema/field_name_collisions/ordinary_PydanticV2BaseModel.py
@@ -1,11 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  ordinary.xsd
-
-from __future__ import annotations
-
-from pydantic import BaseModel, conint
-
-
-class Root(BaseModel):
-    item: conint(ge=-2147483648, le=2147483647)
-    code: str
